Demonstrates that qunit and loglevel are not compatible
When using qunit and loglevel, the QUnit globals for both 'log' and
'module' will prevent loglevel from operating correctly.

This problem has been discussed here:

  http://stackoverflow.com/questions/23770239/qunit-exports-global-functions

It has also been reported here:

  jquery/qunit#228

Although demonstrated with QUnit, it is generally useful for loglevel to
support customizing the logger name and correctly detect the environment
that it is running under.