Add a `setDefaultLevel()` method that sets the log level only if none was previously persisted.
This is useful when initializing scripts; if a developer or user has previously called `setLevel()`, this won’t alter their settings. For example, your application might set the log level to `error` in a production environment, but when debugging an issue, you might call `setLevel("trace")` on the console to see all the logs. If that `error` setting was set using `setDefaultLevel()`, it will still say as `trace` on subsequent page loads and refreshes instead of resetting to `error`, which would have happened if page scripts used `setLevel()`.

Note this also modifies a private method, `loadPersistedLevel()`, to only return whatever valid level has been saved, rather than setting the current level to the saved level. This gives us much more flexibility internally with very little added complexity when calling it.

Fixes #74.