Ensure IE check works in IE < 9
Ok, strictly speaking this isn't too important, as it's been
a long time since IE9 came out, but that's what we say we
support, it's easy enough to support it here, and one of the
big reasons this library exists is to give you reliable
logging, everywhere. It's nice to do that *everywhere*, when
we can.