Add `Logger.rebuild()` method
This implements a new `Logger.rebuild(includeChildren)` method that does the actual work of replacing logging methods on a logger and, optionally, on child loggers. The goal is to make adding plugins (or otherwise changing the method factory) easier by giving users a way to update any child loggers that have already been declared, solving the need for careful declaration ordering as described in #187.

In the ideal case, plugin developers will call this method automatically when their plugin is attached to a logger so users don't have to worry about it at all.