Upgrade Grunt to v1.0.4 (latest v1.0.x)
This upgrades Grunt to v1.0.4, which is the latest of the v1.0.x series
and resolves 6 critical vulnerabilities, including several arbitrary
code execution and command injection vulnerabilities. There are newer
versions of Grunt, but this is the minimum update.

This also required upgrading some related packages that were not
compatible with Grunt 1.x. It also required forking and vendoring
grunt-template-jasmine-requirejs, which is no longer maintained and was
not compatible. I've tried to make the minimum viable changes there and
mark all the changes clearly.

Finally, this required a bunch of minor changes to tests, since the
compatible versions of grunt-contrib-jasmine all use Jasmine 2.x
instead of 1.x, which we were currently on. Unfortunately, this leaves
the Node.js tests on Jasmine 1.x (since they use a different Grunt
plugin). I'll see if it's possible to update that in a follow-on commit.