Upgrade grunt-contrib-qunit to be compatible with Grunt 1.x
The previous version appeared to work fine, but upgrade to grunt-contrib-qunit >= 0.6.0 fixes a peer dependency warning with current versions of Grunt. I've upgraded all the way v2.0.0 here, since that is still compatible with Node.js v0.10.0. v3.0 is also compatible, but switches to headless Chrome, and for now I think it keeps things simpler to have all browser tests using Phantom (but later we should switch to Chrome if we can, since the PhantomJS project is quite dead).