Add basic CI workflow on GitHub Actions
This also breaks up some of the Grunt tasks to match the different jobs in this workflow. My thinking here was to let things that should run in parallel do so, and separate things meant to run in different environments with different requirements.

Browser tests run on Node.js v8 instead of v10 because the grunt plugin we are using relies on a Node.js feature that was removed in v10. This should be fixed in a future update.