Fix async usage in cookies + localstorage tests
Two tests in the cookies + localstorage suite were handling async
operations incorrectly (I think this is a change in Jasmine v2, and I
missed these in #190 because these tests don't run from the command
line). This fixes them, which resolves one test failure. The other is
still failing because of an actual bug in the library.