Fix bug restoring levels from cookies
It turns out loading logger levels from cookies has been broken
(probably for a long time!), but since the tests that run from
`npm test` don't actually test cookies, this was never caught (those
tests use PhantomJS, which does not allow access to cookies from
`document.cookie`, so these tests were skipped). When running tests by
running `grunt integration-test` and then opening the tests in an
actual browser, you can see this test is failing.

The cause is that the cookie value is read starting from beginning of
the whole key-value pair, rather than the start of the cookie, so we'd
read in a value like `"loglevel=INFO"` instead of just `"INFO"`.

Incidentally, fixing this causes *other* tests to fail (uhoh), which I
will address in the next commit.