Remove Coveralls.io tooling
We have not actively posted Coverage data to coveralls.io in several years and retaining the coveralls tooling adds to our maintenance support burden, so this removes it. (If we want to start using it again in the future, they now provide a GitHub Action step, which is probably a better way to do things than with the Grunt plugin we were using.)

This also removes 11 security warnings including 1 that was rated "critical".