Upgrade grunt-contrib-jasmine to v1.2
This is the latest v1.x release, and allows Jasmine browser tests to work in modern Node.js releases (Node.js v12+). The latest version is actually v4, but we can't upgrade to v2+ because there is no compatible version of grunt-template-jasmine-istanbul (top-of-tree there is compatible, but was never released and apparently has unaddressed test failures; developed appears to be stopped, so we also can't expect a future release).