Upgrade grunt-jasmine-node to v0.3.1
This is the latest version and allows grunt-jasmine-node to function in modern versions of Node.js. Unfortunately, it contains a *huge* regression in configuration and requires this super-wonky hack. Some folks in grunt-jasmine-node's issues have proposed other ways to work around this in how you structure your config, but none of those actually work correctly (I did a lot of testing here!).